doc: moved annotations to the near macro documentation. Created backlinks on annotations. Fixed urls
